+++ components/bridge/api/src/main/java/com/flipperdevices/bridge/api/utils/Constants.kt
@@ -26,6 +26,10 @@ object Constants {
         majorVersion = 0,
         minorVersion = 18
     )
+    val API_SUPPORTED_MD5_LISTING = SemVer(
+        majorVersion = 0,
+        minorVersion = 19
+    )
     val API_SUPPORTED_GET_REQUEST = API_SUPPORTED_FLIPPER_ERROR
     val API_SUPPORTED_VERSION = SemVer(majorVersion = 0, minorVersion = 3)
     const val LAGS_FLIPPER_DETECT_TIMEOUT_MS = 10 * 1000L // 10 seconds