+++ components/bridge/rpc/api/src/main/java/com/flipperdevices/bridge/rpc/api/FlipperStorageApi.kt
@@ -1,5 +1,6 @@
 package com.flipperdevices.bridge.rpc.api
 
+import com.flipperdevices.bridge.rpc.api.model.NameWithHash
 import com.flipperdevices.core.progress.ProgressListener
 import java.io.File
 
@@ -9,4 +10,8 @@ interface FlipperStorageApi {
     suspend fun delete(path: String, recursive: Boolean = false)
 
     suspend fun upload(pathOnFlipper: String, fileOnAndroid: File, progressListener: ProgressListener)
+
+    suspend fun listingDirectory(pathOnFlipper: String): List<String>
+
+    suspend fun listingDirectoryWithMd5(pathOnFlipper: String): List<NameWithHash>
 }