+++ components/bridge/rpc/impl/src/main/java/com/flipperdevices/bridge/rpc/impl/api/FlipperStorageApiImpl.kt
@@ -1,11 +1,16 @@
 package com.flipperdevices.bridge.rpc.impl.api
 
 import com.flipperdevices.bridge.api.model.wrapToRequest
+import com.flipperdevices.bridge.api.utils.Constants
 import com.flipperdevices.bridge.rpc.api.FlipperStorageApi
+import com.flipperdevices.bridge.rpc.api.model.NameWithHash
+import com.flipperdevices.bridge.rpc.impl.delegates.FlipperListingDelegate
 import com.flipperdevices.bridge.rpc.impl.delegates.FlipperUploadDelegate
 import com.flipperdevices.bridge.rpc.impl.delegates.MkDirDelegate
 import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
 import com.flipperdevices.core.di.AppGraph
+import com.flipperdevices.core.log.LogTagProvider
+import com.flipperdevices.core.log.info
 import com.flipperdevices.core.progress.ProgressListener
 import com.flipperdevices.protobuf.Flipper
 import com.flipperdevices.protobuf.main
@@ -21,28 +26,32 @@ import javax.inject.Inject
 class FlipperStorageApiImpl @Inject constructor(
     private val flipperServiceProvider: FlipperServiceProvider,
     private val mkDirDelegate: MkDirDelegate,
-    private val flipperUploadDelegate: FlipperUploadDelegate
-) : FlipperStorageApi {
+    private val flipperUploadDelegate: FlipperUploadDelegate,
+    private val listingDelegate: FlipperListingDelegate
+) : FlipperStorageApi, LogTagProvider {
+    override val TAG = "FlipperStorageApi"
+
     override suspend fun mkdirs(path: String) = withContext(Dispatchers.Default) {
         mkDirDelegate.mkdir(flipperServiceProvider.getServiceApi().requestApi, path)
     }
 
-    override suspend fun delete(path: String, recursive: Boolean) = withContext(Dispatchers.Default) {
-        val requestApi = flipperServiceProvider.getServiceApi().requestApi
-        val response = requestApi.request(
-            main {
-                storageDeleteRequest = deleteRequest {
-                    this.path = path
-                    this.recursive = recursive
-                }
-            }.wrapToRequest()
-        ).first()
-        if (response.commandStatus != Flipper.CommandStatus.OK &&
-            response.commandStatus != Flipper.CommandStatus.ERROR_STORAGE_NOT_EXIST
-        ) {
-            error("Failed delete $path with error ${response.commandStatus}")
+    override suspend fun delete(path: String, recursive: Boolean) =
+        withContext(Dispatchers.Default) {
+            val requestApi = flipperServiceProvider.getServiceApi().requestApi
+            val response = requestApi.request(
+                main {
+                    storageDeleteRequest = deleteRequest {
+                        this.path = path
+                        this.recursive = recursive
+                    }
+                }.wrapToRequest()
+            ).first()
+            if (response.commandStatus != Flipper.CommandStatus.OK &&
+                response.commandStatus != Flipper.CommandStatus.ERROR_STORAGE_NOT_EXIST
+            ) {
+                error("Failed delete $path with error ${response.commandStatus}")
+            }
         }
-    }
 
     override suspend fun upload(
         pathOnFlipper: String,
@@ -55,5 +64,30 @@ class FlipperStorageApiImpl @Inject constructor(
             fileOnAndroid = fileOnAndroid,
             externalProgressListener = progressListener
         )
+    }
+
+    override suspend fun listingDirectory(pathOnFlipper: String): List<String> {
+        return listingDelegate.listing(
+            requestApi = flipperServiceProvider.getServiceApi().requestApi,
+            pathOnFlipper = pathOnFlipper
+        )
+    }
+
+    override suspend fun listingDirectoryWithMd5(pathOnFlipper: String): List<NameWithHash> {
+        val serviceApi = flipperServiceProvider.getServiceApi()
+        val md5ListingSupported = serviceApi.flipperVersionApi.isSupported(Constants.API_SUPPORTED_MD5_LISTING)
+        return if (md5ListingSupported) {
+            info { "Use new md5 request api" }
+            listingDelegate.listingWithMd5(
+                requestApi = serviceApi.requestApi,
+                pathOnFlipper = pathOnFlipper
+            )
+        } else {
+            info { "Use deprecated md5 request api" }
+            listingDelegate.listingWithMd5Deprecated(
+                requestApi = serviceApi.requestApi,
+                pathOnFlipper = pathOnFlipper
+            )
+        }
     }
 }