+++ components/faphub/installation/manifest/impl/src/main/java/com/flipperdevices/faphub/installation/manifest/impl/utils/FapManifestsLoader.kt
@@ -3,7 +3,6 @@ package com.flipperdevices.faphub.installation.manifest.impl.utils
 import com.flipperdevices.bridge.api.manager.FlipperRequestApi
 import com.flipperdevices.bridge.api.model.FlipperRequestPriority
 import com.flipperdevices.bridge.api.model.wrapToRequest
-import com.flipperdevices.bridge.dao.api.model.FlipperFileFormat
 import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
 import com.flipperdevices.core.ktx.jre.flatten
 import com.flipperdevices.core.ktx.jre.pmap
@@ -11,20 +10,19 @@ import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
 import com.flipperdevices.faphub.errors.api.throwable.FlipperNotConnected
 import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestConstants.FAP_MANIFESTS_FOLDER_ON_FLIPPER
-import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestConstants.FAP_MANIFEST_EXTENSION
 import com.flipperdevices.faphub.installation.manifest.model.FapManifestItem
 import com.flipperdevices.protobuf.main
-import com.flipperdevices.protobuf.storage.listRequest
 import com.flipperdevices.protobuf.storage.readRequest
 import kotlinx.coroutines.flow.toList
 import java.io.File
-import java.nio.charset.Charset
 import javax.inject.Inject
 
 class FapManifestsLoader @Inject constructor(
     private val flipperServiceProvider: FlipperServiceProvider,
-    private val parser: FapManifestParser
+    private val parser: FapManifestParser,
+    private val cacheLoader: FapManifestCacheLoader
 ) : LogTagProvider {
+
     override val TAG = "FapManifestsLoader"
 
     suspend fun load(): List<FapManifestItem> {
@@ -33,43 +31,28 @@ class FapManifestsLoader @Inject constructor(
             throw FlipperNotConnected()
         }
         info { "Start load manifests" }
-        var manifestNames = getManifestPaths(serviceApi.requestApi)
-        info { "Find ${manifestNames.size} files" }
-        manifestNames = manifestNames.filter { File(it).extension == FAP_MANIFEST_EXTENSION }
-        info { "Find ${manifestNames.size} manifest files" }
-        val fapItems = manifestNames.pmap { name ->
+        val cacheResult = cacheLoader.loadCache()
+        info { "Cache load result is toLoad: ${cacheResult.toLoadNames}, cached: ${cacheResult.cachedNames}" }
+        val fapItemsFromFlipper = cacheResult.toLoadNames.pmap { name ->
             val content = loadManifestFile(
                 requestApi = serviceApi.requestApi,
                 filePath = File(FAP_MANIFESTS_FOLDER_ON_FLIPPER, name).absolutePath
             ) ?: return@pmap null
             parser.parse(content, name)
         }.filterNotNull()
-        info { "Parsed ${fapItems.size} manifests" }
+        info { "Parsed ${fapItemsFromFlipper.size} manifests from flipper" }
+        val fapItemsFromCache = cacheResult.cachedNames.pmap { (file, name) ->
+            parser.parse(file.readBytes(), name)
+        }.filterNotNull()
+        info { "Parsed ${fapItemsFromCache.size} manifests from cache" }
 
-        return fapItems
-    }
-
-    private suspend fun getManifestPaths(requestApi: FlipperRequestApi): List<String> {
-        return requestApi.request(
-            main {
-                storageListRequest = listRequest {
-                    path = FAP_MANIFESTS_FOLDER_ON_FLIPPER
-                }
-            }.wrapToRequest()
-        ).toList().mapNotNull { response ->
-            if (response.hasStorageListResponse()) {
-                response.storageListResponse.fileList
-            } else {
-                info { "Can't find storage list response, $response" }
-                null
-            }
-        }.flatten().map { it.name }
+        return fapItemsFromFlipper + fapItemsFromCache
     }
 
     private suspend fun loadManifestFile(
         requestApi: FlipperRequestApi,
         filePath: String
-    ): FlipperFileFormat? {
+    ): ByteArray? {
         val responseBytes = requestApi.request(
             main {
                 storageReadRequest = readRequest {
@@ -83,8 +66,7 @@ class FapManifestsLoader @Inject constructor(
                 return null
             }
         }.flatten()
-        val textResponse = responseBytes.toString(Charset.forName("UTF-8"))
 
-        return FlipperFileFormat.fromFileContent(textResponse)
+        return responseBytes
     }
 }